New elements replace-partition(-redhat)
replace-partition
-----------------
A standalone element which consumes a base image which was created with
``diskimage-builder`` and rebuilds it without making any packaging changes. This
allows the image contents to be copied to a new block device layout. Use cases
for this element include:

* Rebuilding a whole-disk image with a different partition layout by setting
  ``DIB_BLOCK_DEVICE_CONFIG``
* Rebuilding a whole-disk image with the same partitions but with the sector
  size increased to 4096 bytes

replace-partition-redhat
------------------------

A redhat family specific version of the ``replace-partition`` element.

Change-Id: I7399c4bf6a4d6acfef43f871df0a40e2961ed44e