Reduce LVM extent usage
With 4k disks, it appears that we can encounter cases where
logical block sizes being larger begins to chip away at space
at the beginning *and* end of an LVM Physical Volume (PV).

With GPT, a similar issue occurs, larger logical block sizes
increases the number of bytes used at the beginning and end
of disk for metadata storage.

For the PV, the signature is on the second logical block, and
there is also reference to a potential backup at the end of the
volume. Which means the overall available space goes from just
being impacted by 1kB to at least 8kB. This can then begin to
impact overall alignment which is in 4MB blocks on top of the
base device.

The attempt here is to dial the sizing back just a little bit,
so we avoid assumptions regarding volume sizing ever possibly
conflicting or cases where we somehow end up asking to grow a
volume on a physical volume group which exceeds the number of
extents LVM has calculated.

Example failure this intends to avoid similar errors on a 4k
disk when growvols is triggered:

Exception: Running command failed: cmd "lvextend -L+952065064960B /dev/mapper/vg-lv_thinpool /dev/sdc6"

Which involves Standard Error ouptut along the lines of:

"Insufficient free space: 226990 extents needed, but only 226988 available"

With this change, we will apply logic to *both* the extent size
of the thinpool, if any, and all logical volumes which effectively
reduces the leveraged volumes by 8MB or two LVM extents, because
growvols tries to convert the user's preference of size or percentage
to bytes, which doesn't compensate for the additional lost extents on
a 4k block device.

Change-Id: I303d504f3c822fd534f3e3642d85873ba30d3f68