trivial: provide a little more context for disk size failures
While trying to build a new CI job, I came to realize that the
default image was too small, and the block-device-efi-lvm element
example wants 40GB by default.

Even though that can be easily resolved, it kind of requires taking
a look at the code to understand what is going on. So instead,
we'll now log an error as a hint.

Change-Id: I05bfeae1a2b9aef01f781115fdbd5046c80437f0