Update Nodepool image location in docs
DIB docs refer to the nodepool-builder container image as an easy way to
run disk-image-builder. Previously this image was hosted on docker hub,
but docker made changes to their image hosting policies and all Zuul and
Nodepool images moved to quay.io. Update the docs to refer to the
new/current image location on quay.io.

Change-Id: Ie19e0e828368bfd0c4aab0ac495c45c19aab7b53