Force grub2-install to bypass secureboot complaints.
GRUB_OPTS is already set to --force by default and is being overriden
to the x86_64 target for those architectures. It would be more effective
to append this target to the options instead of overriding it.

In EL9 images, grub2-install will not proceed as it complains about
not supporting secure boot. This isn't appropriate for the way that
diskimage-builder works, and thus --force is needed to bypass this.

Closes-Bug: 1930878
Change-Id: I989bd009548adc7bbf58154577c24980bd62cb2d