Temporarily disable OpenEuler functional testing
After the mirror was updated to a newer version of OpenEuler
packages in Ifcee064959972bf871688baeff7b3288d7517da1 this testing
began to fail looking for the old version, so will need some work to
bring it back into a functioning state. Disable it for now so that
we can merge unrelated changes again.

Change-Id: I40564fb93aed339bc085daf1bd157a5b11218d9a