Change grub variables for style and timeout
This change is done in order to make the troubleshooting of images
in non CI/CD deployments easier. We pick GRUBs upstream defaults to
select as neutral a set of values as possible. If you need additional
time for debugging or hidden menus you should override these values in
your downstream elements.

Change-Id: I08155152b63f629c4eaf72d29ff931e92d7e4b33