CI: Add a tool for displaying CPU flags and QEMU version
As RHEL-10, and related OS projects (CentOS, RockyLinux, and AlmaLinux),
need specific x86_84 machine levels, in this case x86_64-v3, It's helpful
to see which CPUFlags are present in a given node and which version of QEMU
is available.

This adds a small tool to list the CPUFlags needed and those found.  There are
a few tools that will can do similar things but I didn't find one that covered
what we need.

Change-Id: Id8e4f679f6e781dce61d6ace151f3b639dc7edcc