root password for dynamic-login made simpler
Simplifies the definition of root password for dynamic login by lifting
the requirement for a double quote at the end of the command line
argument and letting the user encode the hash value in base64 format.

Compatibility with the legacy behavior is kept but the updated
documentation recomends using base64 as this makes it possible to use
the same command line with syslinux (Bios) or grub (UEFI).

Closes-Bug: #2123878

Change-Id: I3c263746ac2e4e6b1516c0303fb1193dbd9fb480
Signed-off-by: Pierre Crégut <pierre.cregut@orange.com>